requirements.txt
# Web framework
fastapi==0.115.12
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.49
alembic==1.15.2
asyncpg==0.31.0
psycopg2-binary==2.9.10   # синхронный драйвер для Alembic миграций
greenlet==3.3.2             # требуется SQLAlchemy async на Python 3.14

# Validation
pydantic==2.12.5
pydantic[email]
pydantic-settings==2.9.1

# Auth
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.20

# Background tasks
celery==5.5.1
redis==5.2.1
flower==2.0.1

# File storage (Cloudflare R2 / S3)
boto3==1.38.0

# HTTP client (for tests)
httpx==0.28.1

# Testing
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.0
aiosqlite==0.22.1          # in-memory SQLite для тестов
